+++ tensorflow/compiler/plugin/poplar/driver/tools/seed_generator.cc
@@ -30,7 +30,7 @@ void SeedGenerator::PrepareSeedsForReplicas(int64 replication_factor) {
 }
 
 uint64 SeedGenerator::Get(int64 replica_idx) {
-  CHECK(replica_idx < buffer_.size());
+  CHECK(static_cast<size_t>(replica_idx) < buffer_.size());
   return buffer_[replica_idx];
 }
 }  // namespace poplarplugin
